+++ prisma/migrations/20250704111219_disabled_channels/migration.sql
@@ -1,0 +1,2 @@
+-- AlterTable
+ALTER TABLE "Guild" ADD COLUMN     "disabledChannels" TEXT[];