+++ src/commands/settings.ts
@@ -30,6 +30,7 @@ import {
 import { isPassive, setPassive } from "../utils/functions/economy/passive";
 import { createUser, formatNumber, userExists } from "../utils/functions/economy/utils";
 import { setAltPunish } from "../utils/functions/guilds/altpunish";
+import { getDisabledChannels, setDisabledChannels } from "../utils/functions/guilds/channels";
 import { setSlashOnly } from "../utils/functions/guilds/slash";
 import { cleanString } from "../utils/functions/string";
 import { checkPurchases, getEmail, setEmail } from "../utils/functions/users/email";
@@ -115,6 +116,17 @@ cmd.slashData
           .setDescription("automatically punish a user's alts set with $alts when punished")
           .addBooleanOption((option) =>
             option.setName("value").setDescription("yes/no").setRequired(true),
+          ),
+      )
+      .addSubcommand((disableChannels) =>
+        disableChannels
+          .setName("disabled-channels")
+          .setDescription("configure the disabled channels in the server")
+          .addChannelOption((option) =>
+            option
+              .setName("channel")
+              .setDescription("toggle the channel's disabled status")
+              .setRequired(false),
           ),
       ),
   );
@@ -967,6 +979,67 @@ async function run(
       });
   };
 
+  const doDisabledChannels = async () => {
+    let disabledChannels = await getDisabledChannels(message.guild);
+
+    const showChannels = async () => {
+      if (disabledChannels.length === 0) {
+        return send({
+          embeds: [new CustomEmbed(message.member, "there are no disabled channels")],
+        });
+      }
+
+      return send({
+        embeds: [
+          new CustomEmbed(
+            message.member,
+            "disabled channels:\n" + disabledChannels.map((i) => `<#${i}>`).join("\n"),
+          ),
+        ],
+      });
+    };
+
+    let updated = false;
+    for (const channelId of disabledChannels) {
+      if (!message.guild.channels.cache.has(channelId)) {
+        updated = true;
+        await setDisabledChannels(
+          message.guild,
+          disabledChannels.filter((c) => c !== channelId),
+        );
+      }
+    }
+
+    if (updated) {
+      disabledChannels = await getDisabledChannels(message.guild);
+    }
+
+    if (args.length === 2) {
+      return showChannels();
+    }
+
+    const channel = message.mentions.channels.first();
+
+    if (!channel) {
+      return send({ embeds: [new ErrorEmbed("/settings server disabled-channels <channel>")] });
+    }
+
+    if (disabledChannels.includes(channel.id)) {
+      await setDisabledChannels(
+        message.guild,
+        disabledChannels.filter((c) => c !== channel.id),
+      );
+      disabledChannels = await getDisabledChannels(message.guild);
+
+      return showChannels();
+    } else {
+      await setDisabledChannels(message.guild, [...disabledChannels, channel.id]);
+      disabledChannels = await getDisabledChannels(message.guild);
+
+      return showChannels();
+    }
+  };
+
   if (args.length == 0) {
     return send({ embeds: [new CustomEmbed(message.member, "/settings me\n/settings server")] });
   } else if (args[0].toLowerCase() == "me") {
@@ -1000,11 +1073,12 @@ async function run(
   } else if (args[0].toLowerCase() == "server") {
     if (args[1]?.toLowerCase() == "slash-only") {
       return slashOnly();
-    }
-    if (args[1]?.toLowerCase() == "alt-punish") {
+    } else if (args[1]?.toLowerCase() == "alt-punish") {
       return altPunish();
+    } else if (args[1]?.toLowerCase() === "disabled-channels") {
+      return doDisabledChannels();
     } else {
-      const subcommands = ["slash-only", "alt-punish"];
+      const subcommands = ["slash-only", "alt-punish", "disabled-channels"];
       return send({
         embeds: [
           new CustomEmbed(