+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # july 2025
 
 ## changes you'll probably care about
+
 - added better searching and pages to $ach view (@growling_grizzly)
+- you can disable channels in your server /settings server disabled-channels
 
 ## other